fix(ext-node-compat): move worker into extension to fix Deno isolate registry bug
The `upload-extraction-worker` lived in core and dynamically imported
kreuzberg, relying on `tryResolve<NodeCompat>` inside the worker. Deno
Workers run in a separate isolate with their own module state, so the
in-process contract registry is always empty inside the worker —
causing `loadUpload()` on Deno to throw "install ext-node-compat" for
non-text formats even after the extension is registered on the main
thread.

Fix: add `extractInWorker(buffer, mimeType)` to the NodeCompat
contract. Move the worker script into the extension where
`@kreuzberg/wasm` resolves via the extension's own import map, and
have core's `loadUpload()` call the contract method. A synchronous
registry check surfaces the missing-extension error before any worker
is spawned.

Also drops the now-unused worker from core's compile-binary includes.
