+++ extensions/ext-node-compat/src/upload-extraction-worker.ts
@@ -4,12 +4,19 @@
  * Runs `extractBytes` in an isolated Worker thread so that long-running or
  * hung WASM operations cannot block the main server event loop.
  *
- * @module embedding
+ * Lives inside the ext-node-compat extension (not core) so that the
+ * `@kreuzberg/wasm` import and `#kreuzberg-wasm-glue` pre-import resolve
+ * through the extension's own import map. Deno worker isolates do not
+ * share the main-thread NodeCompat contract registry, so the worker
+ * calls into the shared `loadKreuzberg` helper directly rather than
+ * round-tripping through the registry.
+ *
+ * @module extensions/ext-node-compat/upload-extraction-worker
  */
 
 /// <reference lib="deno.worker" />
 
-import { importKreuzberg } from "#veryfront/platform/compat/opaque-deps.ts";
+import { loadKreuzberg } from "./kreuzberg.ts";
 
 interface ExtractRequest {
   buffer: ArrayBuffer;
@@ -24,7 +31,7 @@ interface ExtractResponse {
 self.onmessage = async (event: MessageEvent<ExtractRequest>) => {
   try {
     const { buffer, mimeType } = event.data;
-    const { extractBytes } = await importKreuzberg();
+    const { extractBytes } = await loadKreuzberg();
     const result = await extractBytes(new Uint8Array(buffer), mimeType);
     self.postMessage({ content: result.content } satisfies ExtractResponse);
   } catch (err) {
